% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMiner.R
\docType{package}
\name{DataMiner-package}
\alias{DataMiner-package}
\title{DataMiner: Basic Data Analysis Package}
\description{
A simple package for data analysis and visualization.
}
\author{
\strong{Maintainer}: Your Name \email{your@email.com}

}
\keyword{internal}
