% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMiner.R
\name{format_stats}
\alias{format_stats}
\title{Format statistics as data frame}
\usage{
format_stats(stats_list)
}
\arguments{
\item{stats_list}{Result from desc_stats()}
}
\value{
Formatted data frame
}
\description{
Format statistics as data frame
}
