% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMiner.R
\name{plot_correlations}
\alias{plot_correlations}
\title{Create correlation plot}
\usage{
plot_correlations(data, method = "pearson")
}
\arguments{
\item{data}{Data frame with numeric columns}

\item{method}{Correlation method ("pearson", "kendall", "spearman")}
}
\value{
Heatmap of correlations
}
\description{
Create correlation plot
}
\examples{
plot_correlations(mtcars)
}
