% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMiner.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Create diagnostic plots}
\usage{
plot_distribution(data, col, bins = 30)
}
\arguments{
\item{data}{Input data frame}

\item{col}{Name of numeric column to plot}

\item{bins}{Number of histogram bins (default: 30)}
}
\value{
Combined histogram and boxplot
}
\description{
Create diagnostic plots
}
\examples{
plot_distribution(mtcars, "mpg")
}
